% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prd_pickmtrl_waste.R
\name{prd_pickmtrl_wasteUI_bottom}
\alias{prd_pickmtrl_wasteUI_bottom}
\title{生成区域}
\usage{
prd_pickmtrl_wasteUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
prd_pickmtrl_wasteUI_bottom()
}
