% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prd_pickmtr.R
\name{prd_pickmtrUI}
\alias{prd_pickmtrUI}
\title{prd_pickmtr生成器界面}
\usage{
prd_pickmtrUI(
  tabTitle = "生产领料-倒扣线边库同步",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = prd_pickmtrUI_left,
  func_right = prd_pickmtrUI_right,
  func_bottom = prd_pickmtrUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
prd_pickmtr生成器界面
}
\examples{
prd_pickmtrUI()
}
