% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productWarehousing.R
\name{productWarehousingUI_bottom}
\alias{productWarehousingUI_bottom}
\title{生成区域}
\usage{
productWarehousingUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
productWarehousingUI_bottom()
}
