% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpd_returnstock.R
\name{rpd_returnstockUI}
\alias{rpd_returnstockUI}
\title{rpd_returnstock生成器界面}
\usage{
rpd_returnstockUI(
  tabTitle = "生产退库同步",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = rpd_returnstockUI_left,
  func_right = rpd_returnstockUI_right,
  func_bottom = rpd_returnstockUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
rpd_returnstock生成器界面
}
\examples{
rpd_returnstockUI()
}
