% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saleOut.R
\name{saleOutUI}
\alias{saleOutUI}
\title{saleOut生成器界面}
\usage{
saleOutUI(
  tabTitle = "销售出库",
  colTitles = c("数据同步", "中台查询", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = saleOutUI_left,
  func_right = saleOutUI_right,
  func_bottom = saleOutUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
saleOut生成器界面
}
\examples{
saleOutUI()
}
