% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_miscellaneous_std.R
\name{stk_miscellaneous_stdUI_left}
\alias{stk_miscellaneous_stdUI_left}
\title{请输入文件}
\usage{
stk_miscellaneous_stdUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
stk_miscellaneous_stdUI_left()
}
