% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_miscellaneous_std.R
\name{stk_miscellaneous_stdUI_right}
\alias{stk_miscellaneous_stdUI_right}
\title{生成区域}
\usage{
stk_miscellaneous_stdUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
stk_miscellaneous_stdUI_bottom()
}
