% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_miscellaneous_waste.R
\name{stk_miscellaneous_wasteUI_bottom}
\alias{stk_miscellaneous_wasteUI_bottom}
\title{生成区域}
\usage{
stk_miscellaneous_wasteUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
stk_miscellaneous_wasteUI_bottom()
}
