% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_misdelivery_crush.R
\name{stk_misdelivery_crushUI_right}
\alias{stk_misdelivery_crushUI_right}
\title{生成区域}
\usage{
stk_misdelivery_crushUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
stk_misdelivery_crushUI_bottom()
}
