% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_misdelivery_std.R
\name{stk_misdelivery_stdUI}
\alias{stk_misdelivery_stdUI}
\title{stk_misdelivery_std生成器界面}
\usage{
stk_misdelivery_stdUI(
  tabTitle = "标准其他出库单同步",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = stk_misdelivery_stdUI_left,
  func_right = stk_misdelivery_stdUI_right,
  func_bottom = stk_misdelivery_stdUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
stk_misdelivery_std生成器界面
}
\examples{
stk_misdelivery_stdUI()
}
