% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_stktransferin_material.R
\name{stk_stktransferin_materialUI_left}
\alias{stk_stktransferin_materialUI_left}
\title{请输入文件}
\usage{
stk_stktransferin_materialUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
stk_stktransferin_materialUI_left()
}
