% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_stktransferin_material.R
\name{stk_stktransferin_materialUI_right}
\alias{stk_stktransferin_materialUI_right}
\title{生成区域}
\usage{
stk_stktransferin_materialUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
stk_stktransferin_materialUI_bottom()
}
