% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stk_stktransferin_others.R
\name{stk_stktransferin_othersUI_bottom}
\alias{stk_stktransferin_othersUI_bottom}
\title{生成区域}
\usage{
stk_stktransferin_othersUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
stk_stktransferin_othersUI_bottom()
}
